# backend/src/storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "LMS API Documentation",
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/courses": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get courses",
                "operationId": "courses_get",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Add courses",
                "operationId": "courses_add",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Title",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "subtitle",
                        "in": "query",
                        "description": "Subtitle",
                        "type": "string"
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Description",
                        "type": "string"
                    },
                    {
                        "name": "tag",
                        "in": "query",
                        "description": "Tag",
                        "type": "string"
                    },
                    {
                        "name": "image",
                        "in": "query",
                        "description": "Image",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/admin/courses/{id}": {
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Delete courses",
                "operationId": "courses_delete",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Admin"
                ],
                "summary": "Edit courses",
                "operationId": "courses_edit",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Title",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "subtitle",
                        "in": "query",
                        "description": "Subtitle",
                        "type": "string"
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Description",
                        "type": "string"
                    },
                    {
                        "name": "image",
                        "in": "query",
                        "description": "Image",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "default": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}